requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-countries==7.5.1
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-on-heroku==1.1.2
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-storages==1.14
dj_database_url==2.1.0
docx2pdf==0.1.8
drf-yasg==1.21.7
drf-spectacular==0.26.4
google-apps-meet==0.1.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-cloud-pubsub==2.19.0
Pillow==10.1.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
numpy==1.25.2
langchain==0.0.281
llama-index==0.8.20
openai==0.28.0
pandas==2.1.0
pinecone-client==2.2.2
pyautogen==0.1.10
scikit-learn==1.3.1
tabula-py==2.7.0
tiktoken==0.4.0
unstructured==0.10.12